done register, send otp and verify otp api

+++ uteshop-nodejs/src/services/OTPService.js
@@ -1,10 +1,11 @@
 import crypto from "crypto";
 
 const OTPStore = new Map();
+
 const OTPService = {
-    storeOTP: (email, otp) => {
+    storeOTP: (email, otp, maxAttempts = 5) => {
         const expiry = Date.now() + 5 * 60 * 1000; // 5 minutes from now
-        OTPStore.set(email, { otp, expiry });
+        OTPStore.set(email, { otp, expiry, attemptsLeft: maxAttempts });
     },
 
     generateOTP: () => {
@@ -12,14 +13,28 @@ const OTPService = {
     },
 
     verifyOTP: (email, otp) => {
+        console.log(email + " " + otp);
         const record = OTPStore.get(email);
-        if (!record) return false;
+        console.log(record);
+        console.log(OTPStore);
+        if (!record) return {success: false, message: "OTP expired or not existed, try requesting another OTP"};;
         if (record.expiry < Date.now()) {
             OTPStore.delete(email);
-            return false;
+            return {success: false, message: "OTP expired"};
         }
 
-        return record.otp === otp ? (OTPStore.delete(email), true) : false;
+        if (record.otp.attemptsLeft <= 0) {
+            OTPStore.delete(email);
+            return {success: false, message: "Maximum attempts of this OTP reached"};
+        }
+
+        if (otp !== record.otp) {
+            record.otp.attemptsLeft--;
+            return {success: false, message: `{Wrong OTP, try again. You still have ${record.otp.attemptsLeft} attempts left}`}
+        }
+        
+        OTPStore.delete(email);
+        return {success: true, message: "Verify successfully"};
     }
 }
